requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bleach
blinker
bokeh
certifi
charset-normalizer
click
colorama
contourpy
distro
Flask
frozenlist
h11
h2
hpack
httpcore
httpx
Hypercorn
hyperframe
idna
itsdangerous
Jinja2
jiter
linkify-it-py
lxml
Markdown
markdown-it-py
MarkupSafe
mdit-py-plugins
mdurl
multidict
narwhals
numpy
openai
packaging
pandas
panel
param
pillow
priority
propcache
pydantic
pydantic_core
Pygments
PyPDF2
python-dateutil
python-docx
python-dotenv
pytz
pyviz_comms
PyYAML
Quart
redlines
requests
rich
rich-click
six
sniffio
tornado
tqdm
typing-inspection
typing_extensions
tzdata
uc-micro-py
urllib3
webencodings
Werkzeug
wsproto
xyzservices
yarl